requirements.txt
aiohttp
beautifulsoup4
aiofiles
tqdm
python-dateutil
google-genai>=0.16.0
hnswlib
uvicorn
fastapi
pydantic
orjson
requests
numpy
sentence-transformers
python-dotenv
playwright
einops
cross-encoder
playwright
